requirements.txt
google-generativeai
google-api-python-client
yt-dlp
pandas
python-dotenv
matplotlib
seaborn
jinja2
pathlib